requirements.txt
pandas==2.1.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
pyyaml==6.0.1
dbt-core==1.11.8
dbt-postgres==1.10.0
